Chatbot/requirements2.txt
altair==5.4.1
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
faiss-cpu==1.7.4
filelock==3.16.1
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.42.1
networkx==3.1
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
protobuf==5.29.5
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.19.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
rich==13.9.4
rpds-py==0.20.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
six==1.17.0
smmap==5.0.2
streamlit==1.40.1
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
zipp==3.20.2
